# [aiohttp] - use lcg as random generator