--- a/frameworks/Python/aiohttp/aiohttp.dockerfile
+++ b/frameworks/Python/aiohttp/aiohttp.dockerfile
@@ -4,7 +4,9 @@ ADD ./ /aiohttp
 
 WORKDIR /aiohttp
 
-RUN pip3 install -r /aiohttp/requirements-cpython.txt
+RUN pip3 install -r /aiohttp/requirements-cpython.txt && \
+    pip3 install cython==3.1.2 setuptools==80.9.0 && \
+    python3 setup.py build_ext --inplace
 
 ENV CONNECTION=RAW