--- a/frameworks/Python/aiohttp/setup.py
+++ b/frameworks/Python/aiohttp/setup.py
@@ -1,0 +1,12 @@
+from setuptools import setup, Extension
+from Cython.Build import cythonize
+
+setup(
+    ext_modules=cythonize([
+        Extension(
+            "app.random_utils",
+            ["app/random_utils.pyx"],
+            extra_compile_args=["-O3"]
+        ),
+    ])
+)